Mark `WoltLabSuite/Core/Ui/Message/Quote` as deprecated
