+++ wcfsetup/install/files/style/ui/tooltip.scss
@@ -26,7 +26,7 @@
 	&.interactive {
 		pointer-events: all;
 
-		> span {
+		> span, button {
 			cursor: pointer;
 
 			&:not(:first-child) {
